% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{sim_data}
\alias{sim_data}
\title{Simulate survival data from the stratified Cox model with time-varying coefficients.}
\usage{
sim_data(nF = 20, p = 10, seed = 1)
}
\arguments{
\item{nF}{Number of strata.}

\item{p}{Number of covariates.}

\item{seed}{Random seed.}
}
\value{
A list containing the stratified survival data.
\describe{
\item{delta}{event indicator}
\item{time}{survival time}
\item{facility}{stratum ID}
\item{z}{covariates}
}
}
\description{
Simulate survival data from the stratified Cox model with time-varying coefficients.
}
\examples{
dat1 = sim_data(nF = 20, p = 10, seed = 1)
}
