% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{sim_data_fun}
\alias{sim_data_fun}
\title{Simulate survival data from the stratified Cox model with time-varying coefficients.}
\usage{
sim_data_fun(
  nF = 50,
  alph = 50,
  seed = 12,
  censor.max = 30,
  end.time = 3,
  fun.list = fun.list,
  z.type = z.type,
  z.para = z.para
)
}
\arguments{
\item{nF}{Number of strata.}

\item{alph}{Average size of strata.}

\item{seed}{Random seed.}

\item{censor.max}{Maximum value in the censoring time distribution.}

\item{end.time}{Censoring time.}

\item{fun.list}{A list of coefficient functions.}

\item{z.type}{Covariance type: 'Ind' (independent), 'AR(1)' (AR(1)), and 'CS' (compound symmetry)}

\item{z.para}{Correlation coefficient for covariance matrix.}
}
\value{
A list containing the stratified survival data.
\describe{
\item{delta}{event indicator}
\item{time}{survival time}
\item{facility}{stratum ID}
\item{z}{covariates}
}
}
\description{
Simulate survival data from the stratified Cox model with time-varying coefficients.
}
\examples{
z.type = 'Ind' # 'Ind', 'AR1', 'CS'
z.para = 0.5
f1<-function(x){
 (-x^2+3)
}
f1 =  Vectorize(f1)
f2<-function(x){
 2*log(x+0.01)
 }
f2 = Vectorize(f2)
f3<-function(x){
  2*(-3/(x+1)+1)
}
f3 = Vectorize(f3)
fun.list=list(f1,f2,f3)
dat = sim_data_fun(nF = 50, alph=50,seed=12,censor.max=30,end.time=3,fun.list=fun.list, z.type = z.type, z.para = z.para)
}
