% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{stra_sampling}
\alias{stra_sampling}
\title{Sampling within stratification.}
\usage{
stra_sampling(size, prop)
}
\arguments{
\item{size}{A vector of sizes of each stratum.}

\item{prop}{Proportion of sample size to original size.}
}
\value{
A vector of sample indices.
}
\description{
Sampling within stratification.
}
\examples{
stra_sampling(c(3,5,3), prop = 1)
}
