% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{TimeVarying_NR}
\alias{TimeVarying_NR}
\title{Fit stratified time-varying Cox model using Newton-Raphson algorithm.}
\usage{
TimeVarying_NR(
  time,
  delta,
  z,
  facility = NULL,
  knot = 5,
  M_stop = 500,
  rate = 0.001,
  tol = 10^(-6)
)
}
\arguments{
\item{time}{Survival time.}

\item{delta}{Event indicator.}

\item{z}{Covariates.}

\item{facility}{Stratum ID.}

\item{knot}{Number of B spline bases.}

\item{M_stop}{Maximum number of iteration steps.}

\item{rate}{Step size.}

\item{tol}{Convergen criterion.}
}
\value{
A list of fitting results.
\describe{
\item{theta}{Estimated B spline coefficients}
\item{beta}{Estimated time-varying coefficient functions}
\item{llk_all}{Fitted partial likelihood in each step}
\item{likelihood_stratify}{likelihood_stratify}
\item{pvalue}{P values for testing constant coefficients}
\item{basis}{B spline bases}
}
}
\description{
Fit stratified time-varying Cox model using Newton-Raphson algorithm.
}
\examples{
\dontrun{
delta = dat1$delta
facility = dat1$facility
z = dat1$z
time = dat1$time
#fit_NR = TimeVarying_NR(time, delta, z, facility, knot = 6, M_stop = 500, rate = 0.001, tol = 10^(-6))
}
}
